Extract abstract class `ControllerPropertyDefinition`
