+++ src/controller_definition.ts
@@ -2,7 +2,7 @@ import { Project } from "./project"
 import { ParseError } from "./parse_error"
 
 import { identifierForContextKey } from "@hotwired/stimulus-webpack-helpers"
-import { SourceLocation } from "acorn"
+import { MethodDefinition, ValueDefinition, ClassDefinition, TargetDefinition } from "./controller_property_definition"
 
 type ParentController = {
   controllerFile?: string
@@ -14,33 +14,6 @@ type ParentController = {
   type: "default" | "application" | "package" | "import" | "unknown"
 }
 
-export class Definition {
-  constructor(
-    public readonly name: string,
-    public readonly loc?: SourceLocation,
-    public readonly definitionType: "decorator" | "static" = "decorator",
-  ) {}
-}
-
-type ValueDefinitionValue = Array<any> | boolean | number | object | string | undefined
-export class ValueDefinition extends Definition {
-  constructor(
-    name: string,
-    public readonly valueDef: { type: string; default: ValueDefinitionValue },
-    loc?: SourceLocation,
-    definitionType: "decorator" | "static" = "decorator",
-  ) {
-    super(name, loc, definitionType)
-  }
-
-  public static defaultValuesForType = {
-    Array: [],
-    Boolean: false,
-    Number: 0,
-    Object: {},
-    String: "",
-  } as { [key: string]: ValueDefinitionValue }
-}
 
 export class ControllerDefinition {
   readonly path: string
@@ -50,15 +23,12 @@ export class ControllerDefinition {
   isTyped: boolean = false
   anyDecorator: boolean = false
 
-  readonly _methods: Array<Definition> = []
-  readonly _targets: Array<Definition> = []
-  readonly _classes: Array<Definition> = []
+  readonly _methods: Array<MethodDefinition> = []
+  readonly _targets: Array<TargetDefinition> = []
+  readonly _classes: Array<ClassDefinition> = []
   readonly _values: { [key: string]: ValueDefinition } = {}
 
   readonly errors: ParseError[] = []
-  get hasErrors() {
-    return this.errors.length > 0
-  }
 
   static controllerPathForIdentifier(identifier: string, fileending: string = "js"): string {
     const path = identifier.replace(/--/g, "/").replace(/-/g, "_")
@@ -71,22 +41,22 @@ export class ControllerDefinition {
     this.path = path
   }
 
-  // getters for converting internal representations to not break the API
+  get hasErrors() {
+    return this.errors.length > 0
+  }
+
   get methods() {
     return this._methods.map((method) => method.name)
   }
 
-  // getters for converting internal representations to not break the API
   get targets() {
     return this._targets.map((method) => method.name)
   }
 
-  // getters for converting internal representations to not break the API
   get classes() {
     return this._classes.map((method) => method.name)
   }
 
-  // getters for converting internal representations to not break the API
   get values() {
     return Object.fromEntries(Object.entries(this._values).map(([key, def]) => [key, def.valueDef]))
   }
